Run packet loss congestion control algorithm on receiver feedback
As specified in GCC we should run the packet loss estimation algorithm
when we receive the RTCP feedback.

This also starts making use of the feedback packet loss computation
instead of twcc one as it is more accurate since it should take into
account reordering. We aggregate the packet loss from the difference
sources weighting it with the number packets that have been tried to be
sent over the network since the last feedback.